# backend/institutotcnicorical-1ec-Pergola-1.0.0-swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/institutotcnicorical-1ec/Pergola/1.0.0"
  } ],
  "info" : {
    "description" : "EL API de Pérgola se encarga de manejar la base de datos de la empresa de Joyería Pérgola.",
    "version" : "1.0.0",
    "title" : "Pergola API",
    "contact" : {
      "email" : "you@your-company.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths" : {
...
    "/api/collections" : {
      "get" : {
        "summary" : "GET COLLECTIONS",
        "tags" : [ "Collections" ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "operationId" : "get-api-collections",
        "description" : "Listar todas las colecciones."
      },
      "post" : {
        "summary" : "POST COLLECTIONS",
        "tags" : [ "Collections" ],
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "operationId" : "post-api-collections",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "name", "description", "isActive" ],
                "properties" : {
                  "name" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "nqygqmwf1fcod"
                    }
                  },
                  "description" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "qy7bfwy89tuy9"
                    }
                  },
                  "isActive" : {
                    "type" : "boolean",
                    "x-stoplight" : {
                      "id" : "rhdpstit7culj"
                    }
                  }
                }
              }
            }
          }
        },
        "description" : "Crear una colección (con imagen)."
      }
    },
    "/api/collections/:id" : {
      "get" : {
        "summary" : "GET COLLECTIONS - :ID",
        "tags" : [ "Collections" ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "operationId" : "get-api-collections-:id",
        "parameters" : [ {
          "schema" : {
            "type" : "string"
          },
          "in" : "query",
          "name" : "id",
          "required" : true
        } ],
        "description" : "Obtener una colección por ID."
      },
      "put" : {
        "summary" : "PUT COLLECTIONS",
        "tags" : [ "Collections" ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "operationId" : "put-api-collections-:id",
        "parameters" : [ {
          "schema" : {
            "type" : "string"
          },
          "in" : "query",
          "name" : "id",
          "required" : true
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "name", "description", "isActive" ],
                "properties" : {
                  "name" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "phl45ppev1n4c"
                    }
                  },
                  "description" : {
                    "type" : "string",
                    "x-stoplight" : {
                      "id" : "bu48oqg25k1vf"
                    }
                  },
                  "isActive" : {
                    "type" : "boolean",
                    "x-stoplight" : {
                      "id" : "uk4iz1iubbgq4"
                    }
                  }
                }
              }
            }
          }
        },
        "description" : "Actualizar colección (con imagen)."
      },
      "delete" : {
        "summary" : "DELETE COLLECTIONS",
        "tags" : [ "Collections" ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "operationId" : "delete-api-collections-:id",
        "parameters" : [ {
          "schema" : {
            "type" : "string"
          },
          "in" : "query",
          "name" : "id",
          "required" : true
        } ],
        "description" : "Eliminar colección."
      }
    },
...
  },
...
}